pyproject.toml
[tool.poetry]
name = "news_extractor_engine"
version = "1.0.0"
authors = ["Avin Divakara <divakaraavin@gmail.com>"]
description = "A Python package for extracting news from various sources."


[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.2.2"
bs4 = "^0.0.2"
aiohttp = "^3.9.5"
feedparser = "^6.0.11"
python-dotenv = "^1.0.1"
polars = "^1.1.0"
discord-webhook = "^1.3.1"
markdownify = "^0.13.1"
fastapi = "^0.111.1"
pymongo = "^4.8.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"